@@ -193,7 +193,7 @@
         "@babel/core": "^7.26.10",
         "@cypress/vite-dev-server": "^6.0.3",
         "@etchteam/storybook-addon-status": "^5.0.0",
-        "@frontify/eslint-config-react": "^0.17.7",
+        "@frontify/eslint-config-react": "^1.0.3",
         "@frontify/fondue-icons": "workspace:^",
         "@storybook/addon-a11y": "^8.6.12",
         "@storybook/addon-actions": "^8.6.12",
@@ -225,8 +225,7 @@
         "cypress": "^14.3.0",
         "cypress-real-events": "^1.14.0",
         "esbuild": "^0.25.2",
-        "eslint": "^8.57.1",
-        "eslint-plugin-unicorn": "56.0.1",
+        "eslint": "^9.33.0",
         "eslint-plugin-notice": "^1.0.0",
         "eslint-plugin-storybook": "^0.12.0",
         "fast-glob": "^3.3.3",
