@@ -104,7 +104,7 @@ describe('Popover Component', () => {
     });
 
     it('should render the popover with arrow', () => {
-        cy.mount(<PopoverComponent withArrow={true} />);
+        cy.mount(<PopoverComponent withArrow />);
         cy.get(POPOVER_TRIGGER).click();
         cy.get(POPOVER_ARROW).should('exist');
     });
