@@ -230,7 +230,7 @@ const TextareaTemplate: StoryFn<TextareaProps> = (args) => {
     }, [value, setInput]);
 
     return (
-        <Box className="tw-w-[80%] tw-ml-auto tw-mr-auto">
+        <Box className="tw-w-4/5 tw-ml-auto tw-mr-auto">
             <Textarea {...args} onChange={setInput} value={input} />
         </Box>
     );
@@ -294,7 +294,7 @@ export const WithFormFieldCombination: StoryFn<TextareaProps> = (args) => {
     }, [input, handleValidation]);
 
     return (
-        <Box className="tw-w-[80%] tw-ml-auto tw-mr-auto">
+        <Box className="tw-w-4/5 tw-ml-auto tw-mr-auto">
             <FormField
                 label={{
                     text: 'Label from the FormField',
