@@ -11,7 +11,7 @@ const TOAST_ID = '[data-test-id=toast]';
 
 describe('Toast Rendering', () => {
     it('should render toast', () => {
-        cy.mount(<Toast isOpen={true} style={ToastStyle.Loud} icon={<IconLinkBox24 />} />);
+        cy.mount(<Toast isOpen style={ToastStyle.Loud} icon={<IconLinkBox24 />} />);
         cy.get(TOAST_ID).should('exist');
     });
 
@@ -22,7 +22,7 @@ describe('Toast Rendering', () => {
 
     it('should render toast with correct children', () => {
         cy.mount(
-            <Toast isOpen={true} style={ToastStyle.Loud} icon={<IconLinkBox24 />}>
+            <Toast isOpen style={ToastStyle.Loud} icon={<IconLinkBox24 />}>
                 Toast
             </Toast>,
         );
@@ -30,7 +30,7 @@ describe('Toast Rendering', () => {
     });
 
     it('should render toast with correct icon', () => {
-        cy.mount(<Toast isOpen={true} style={ToastStyle.Loud} icon={<IconLinkBox24 />} />);
+        cy.mount(<Toast isOpen style={ToastStyle.Loud} icon={<IconLinkBox24 />} />);
         cy.get(TOAST_ID)
             .find('svg')
             .should(($svg) => {
@@ -40,7 +40,7 @@ describe('Toast Rendering', () => {
 
     it('should render toast with the correct background color', () => {
         const style = ToastStyle.Loud;
-        cy.mount(<Toast isOpen={true} style={style} icon={<IconLinkBox24 />} />);
+        cy.mount(<Toast isOpen style={style} icon={<IconLinkBox24 />} />);
         cy.get(TOAST_ID)
             .children()
             .first()
@@ -51,14 +51,14 @@ describe('Toast Rendering', () => {
 
     it('should render a dark themed version of the toast', () => {
         const style = ToastStyle.Loud;
-        cy.mount(<Toast isOpen={true} theme="dark" style={style} icon={<IconLinkBox24 />} />);
+        cy.mount(<Toast isOpen theme="dark" style={style} icon={<IconLinkBox24 />} />);
         cy.get(TOAST_ID).should('have.class', 'tw-bg-box-neutral-mighty');
     });
 
     it('should be able to use the keyboard for navigation within the toast', () => {
         const onClickStub = cy.stub();
         cy.mount(
-            <Toast isOpen={true} style={ToastStyle.Loud} icon={<IconLinkBox24 />}>
+            <Toast isOpen style={ToastStyle.Loud} icon={<IconLinkBox24 />}>
                 <Button onClick={onClickStub}>I am focusable</Button>
             </Toast>,
         );
